ci: Do not build Docker images on every PR
The images are built whenever we merge to `main`. Building the E2E node,
especially, takes a substantial amount of time and should ideally only
happen as we merge to `main`.

We've also proven that these builds do work now, so there's no need to
continue building these images on every push to every PR.

Signed-off-by: Thane Thomson <connect@thanethomson.com>