consensus: rename (*PeerState).ToJSON to MarshalJSON
In (*PeerState).PickSendVote, there is a Debug-level log that includes
the PeerState value as a logging field. By default, zerolog json-encodes
a struct passed as a log field (when the struct doesn't implement
zerolog.LogObjectMarshaler).

Because PeerState didn't have a MarshalJSON method, the JSON encoder
fell back to reflection to encode the PeerState value. Reflection did
not acquire the lock, and there were data races resulting from an
unsynchronized read while logging the PeerState, and concurrent (locked)
writes at least during (*PeerState).SetHasProposal and
(*PeerState).SetHasVote.

Given that there was only one call to (*PeerState).ToJSON in the
cometbft repo, it seemed appropriate to just rename ToJSON to
MarshalJSON, as opposed to leaving ToJSON for backwards compatibility.
Any third party calls to ToJSON should be able to easily change the
method being called.