light client: Address panic in response to empty signed header (#577)
Closes #575.

This at least addresses the panic in the light client, instead turning it into a temporary error (i.e. that the node does not yet have the signed header for the requisite height, which is true here).

I've tested this patch on the `v0.34.x` branch and it prevents the panic, but instead causes the E2E tests to hang indefinitely. As such, I've logged #576.

The code change is identical across `main`, `v0.37.x` and `v0.34.x`, so backports should be clean. I'll add changelog entries on the backport branches prior to merging them.

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use [unclog](https://github.com/informalsystems/unclog) to manage our changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code comments