Rephrase log when committed TX not in local mempool + make it debug (#738)
Closes #737

If the execution hits the log line changed in this PR, it doesn't mean there is an error condition.
A TX included in a block does not need to be present in the local mempool of a nodes at the time it is processing the new block.
As the info is still useful, we re-word the message and turn it into a Debug log

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use [unclog](https://github.com/informalsystems/unclog) to manage our changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code comments

(cherry picked from commit 6498d67efdf0a539e3ca0dc3e4a5d7cb79878bb2)