[STAB-17] Make unsynchronized local client (#3)
This is effectively a copy of the fully unsynchronized local client found in https://github.com/tendermint/tendermint/pull/9660. Note that this was reverted and replaced by a version that has a mutex per instance of the client in https://github.com/tendermint/tendermint/pull/9830. This change can be removed once a fully unsynchronized client is added back.