proto: abci/v4 package with renamed enums
Rename all enum types and their value names defined in abci
to correspond to buf naming guidelines: no enum definitions nested
in message definitions, value names prefixed with their enum type name,
the zero value has suffix _UNDEFINED.
The only exception to the latter is CHECK_TX_TYPE_NEW that has
to be grandfathered in because of the semantics of the wire value.

Provide versioned definitions of all types that included the enums.