Extend ABCI `max_block_size` parameter to give extended control to the app (#1003)
Closes #980

This PR is the result of internal discussions (last year, but also last week), and interactions with the Vega team in order to address their use case.

The gist of the agreed solution is to interpret value -1 for `ConsensusParams.Block.MaxBytes` as the Application wanting to enforce a size limit at its level, via `PrepareProposal` and `ProcessProposal`. Additionally, it instructs CometBFT to send all transactions in the mempool upon calling `PrepareProposal` so that the application can apply its own transaction ordering policies to *all* transactions known to the node.

This solution is also the most _surgical_ among all candidate solutions in that the least amount of logic/spec needed to be added/removed/modified.

So, at CometBFT level, a value of -1 for `ConsensusParams.Block.MaxBytes` implies:

* The limit enforced at CometBFT level is the generic `types.MaxBlockSizeBytes` (hard-coded to 100MB)
* CometBFT will reap all transactions from the mempool when creating a proposed block (just before calling `PrepareProposal`)

---

#### PR checklist

- [x] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use [unclog](https://github.com/informalsystems/unclog) to manage our changelog)
- [x] Updated relevant documentation (`docs/` or `spec/`) and code comments

(cherry picked from commit c138e785c992b3a76cb25fdf2e0820e29c56c1af)