config: Refactor to match updated ADR
ADR-101 was updated recently to reflect the desired configuration file
changes - this updates the code to match the ADR.

It also changes the "frequency" variable to more accurately be called an
"interval", since it represents a sleep period between pruning
operations as opposed to the number of times it should be run per
second. The interval is also defined as a time.Duration now to
facilitate more ergonomic usage from the configuration file.

Signed-off-by: Thane Thomson <connect@thanethomson.com>