state: Make pruner belong to Node primarily
Move the pruner "ownership" the Node. This doesn't actually have a
meaningful effect on the lifetime of the pruner, but it is more of a
signal to readers of the code that it should be considered a service in
its own right, and the block executor simply has access to it to trigger
one method call.

Signed-off-by: Thane Thomson <connect@thanethomson.com>