grpc: Refactor block service client to avoid panic on channel close
When the caller controls the channel creation, it appears as though
there is still a non-deterministic occurrence of panicking when the
client attempts to send a message via the result channel.

This refactors the interface such that the block service's
GetLatestHeight method returns a channel whose lifecycle it controls,
but still offers the caller the ability to control channel size through
options.

Signed-off-by: Thane Thomson <connect@thanethomson.com>