Adapted the logic for setting the retain height. The retain height is not accepted only if it is less than the current min height.
Fixed potential corner case in checking for the minimum height in case no retain height is set.
This corner case should not be present in practice as the app retain height is expected to be set
but it is a safety check in case the logic of setting it on startup changes down the road.