node: verify genesis doc hash against the file contents rather than remarshalled JSON (#1293)
* node: verify genesis doc hash against the file

Marshalling the doc data back into JSON to checksum is wasteful.
In the future, the hash of the genesis file may be passed as a CLI or
config parameter to help control coordinated upgrades. We want that
hash to be the same as the hash stored in the database.

* node: centralize genesis doc hash verification

Don't put the requirement to verify the JSON content checksum
onto every genesis doc provider, do it in one place in the main code.

* node: link the issue on the recent FIXME comment

* Changelog entry for #1293