test: respect P2PConfig fuzzing configuration in MultiplexTransport (#1414)
As part of porting the cosmos-sdk simulator to running on top of
its testutil/network runner, we need a way to simulate adversarial
network connections between validators. This change adds the
existing fuzzing configuration of P2PConfig to MConnConfig and uses it
to conditionally create an unreliable connection in MultiplexTransport.

Co-authored-by: Elias Naur <mail@eliasnaur.com>
Co-authored-by: Sergio Mena <sergio@informal.systems>