ADR 101: Refactor height check-related logic and tests (#1271) (#1473)
* state: Invert logic of whether retain heights are set to be more readable



* state: Refactor findMinRetainHeight logic for clarity and to respect config



* state: Rename findMinRetainHeight to findMinBlockRetainHeight for clarity



* state: Rename Pruner.SetApplicationRetainHeight to SetApplicationBlockRetainHeight for clarity



* state: Rename pruner SetCompanionRetainHeight to SetCompanionBlockRetainHeight for clarity



* state: Refactor - shorten local variable name



* state: Refactor - rename local method name for clarity



* state: Refactor - remove redundant condition



* state: Refactor - use helper instead of redefining logic



* state: Refactor - shorten local variable names



* state: Refactor - shorten local variable names



* state: Refactor Pruner.SetApplicationBlockRetainHeight logic

The application retain height should be set regardless of what the
companion retain height is. Pruning should take place based on whichever
of the two values is lower. There is no need to consider the companion
retain height in this logic.



* state: Refactor Pruner.SetCompanionBlockRetainHeight logic

Similar refactor to that done for
Pruner.SetApplicationBlockRetainHeight. There is no need to consider the
application retain height during this function call.



* state: Refactor Pruner.SetABCIResRetainHeight logic

Aligns the logic of this method similar to that in the other retain
height setter methods. Also fixes a logic bug where setting the retain
height would skip updating the metrics.



* state: Refactor - shorten local variable names



* state: Simplify pruner logic assuming retain heights are always set

We can simplify the pruner logic if we assume that the initial retain
heights are always set on node startup.



* node: Tell pruner whether companion is enabled



* state: Expand error message detail



* state+store: Align tests with new logic



* state: Refactor - shorten local variable names



* rpc/grpc: Return trace IDs in error responses from pruning service



* node: Refactor/expand pruning service initialization logic



* test/e2e: Minor cosmetic tweaks to gRPC tests



* test/e2e: Enable data companion pruning



* node: Refactor - extract pruner creation as function



* test/e2e: Only enable companion-based pruning on full nodes and validators



* state: Only start ABCI results pruning routine when data companion is enabled



* state: Fix TestMinRetainHeight logic



* node: Fix companion retain height initialization logic



* test/e2e: Expand testing framework

Allows explicit control over whether or not pruning should take a data
companion into account for full nodes or validators in our E2E
framework.

This allows greater flexibility in the E2E framework to test nodes both
with and without data companion-related functionality.

Expands our standard CI manifest to expect data companions attached to
two of the nodes.



* state: Rename background routines



* node: Apply change from code review



---------

Signed-off-by: Thane Thomson <connect@thanethomson.com>
Co-authored-by: Thane Thomson <connect@thanethomson.com>