docs: remove check from PrepareProposal
This check only confuses readers. PrepareProposal returns
`proposal.Txs` (original list), not `txs`. Thus for loop is essentially
noop and can be removed. Even if we modify PrepareProposal to return
txs, it confuses application devs since CometBFT respects the limit when
creating a proposal! Therefore, application devs should only check the
limit when modifying the transactions, which is not the case here.