PBTS: migrating to `main` the new version of the specification (#1973)
This PR cherry-picks a number of commits present in the `master` branch
of this repository.

Some of the commits are from the old `tendermint/spec` repository,
namely:

- tendermint/spec#369
- tendermint/spec#373
- tendermint/spec#375
- tendermint/spec#398
- tendermint/spec#399

This content has then been migrated to the `tendermint/tendermint`
repository, from which follow PRs were considered:

- tendermint/tendermint#7804 (**not cherry-picked**)
- tendermint/tendermint#8018
- tendermint/tendermint#8096

Most of the conflicts and the fixes needed refer to links to content,
that have been updated or referred to old repositories.

The MD files that are currently under
`spec/consensus/proposer-based-timestamp` were moved to the `v1` subdir.
Files with similar names, but the `-002` suffix were added with the new
solution, replacing the old ones.

The differences from the original content, on `master` and the content
of this PR are minimal, as one can check by running

` git diff master spec/consensus/proposer-based-timestamp/`

Finally, the entry point of the added content can be found
[here](https://github.com/cometbft/cometbft/blob/cason/pbts-spec/spec/consensus/proposer-based-timestamp/README.md).

---------

Co-authored-by: Daniel <daniel.cason@usi.ch>
Co-authored-by: Josef Widder <44643235+josef-widder@users.noreply.github.com>
Co-authored-by: William Banfield <4561443+williambanfield@users.noreply.github.com>
Co-authored-by: Kukovec <jure.kukovec@gmail.com>
Co-authored-by: Daniel Cason <cason@gandria>
Co-authored-by: M. J. Fromberger <fromberger@interchain.io>
Co-authored-by: Anton Kaliaev <anton.kalyaev@gmail.com>
(cherry picked from commit 84339efe84260ec1cbd125db76b01e9478a30326)