Forward port of tendermint/tendermint#7605, tendermint/spec#393, and tendermint/tendermint#8142 (#2018)
Closes #2003

This is a pure forward-port of tendermint/tendermint#7605, which was a
"squashed merge" of several other PRs.

There are several refactorings included in tendermint/tendermint#7605
that have been left out of the cherry-pick (not necessary and their
absence makes this PR smaller):
- tendermint/tendermint#6969
- Everything that has to do with context threading, and `RandState` vs.
`MakeState` (since the logic this changes is anyway not present on
`main` anyway)

Finally, two further commits had to be be forward-ported together with
tendermint/tendermint#7605:
- tendermint/spec#393 When PBTS was implemented (end of 2021 - beginning
of 2022) the protobufs were held in a different repo (spec). This PR on
that repo contains those changes, which are needed by
tendermint/tendermint#7605
- tendermint/tendermint#8142 Minor adjustments to Synchrony parameters
that was committed separately on `master`

---

#### Original description of tendermint/tendermint#7605:

This pull request merges in the changes for implementing Proposer-based
timestamps into `master`. The power was primarily being done in the
`wb/proposer-based-timestamps` branch, with changes being merged into
that branch during development. This pull request represents an
amalgamation of the changes made into that development branch. All of
the changes that were placed into that branch have been cleanly rebased
on top of the latest `master`. The changes compile and the tests pass
insofar as our tests in general pass.

These changes have been extensively reviewed during development. There
is not much new here. In the interest of making effective use of time, I
would recommend against trying to perform a complete audit of the
changes presented and instead examine for mistakes that may have
occurred during the process of rebasing the changes. I gave the complete
change set a first pass for any issues, but additional eyes would be
very appreciated.

In sum, this change set does the following:
closes #6942
merges in #6849

---

#### PR checklist

- [x] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [x] Updated relevant documentation (`docs/` or `spec/`) and code
comments

---------

Co-authored-by: William Banfield <4561443+williambanfield@users.noreply.github.com>
Co-authored-by: Daniel <daniel.cason@informal.systems>
Co-authored-by: glnro <8335464+glnro@users.noreply.github.com>