PBTS: forward port of tendermint/tendermint#7709 (#2089)
Part of #1731.

The original PR precedes tendermint/tendermint#8488, a PR that adopted
`metricsgen` for metric generation. This PR was forward ported to this
repository as in tendermint/tendermint#9178. Some changes were added to
the forward-ported PR to match the behaviour and description of the
added metric with this follow-up change.

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

---------

Co-authored-by: William Banfield <4561443+williambanfield@users.noreply.github.com>
Co-authored-by: Sergio Mena <sergio@informal.systems>