test(localnet): Add monitoring tools for localnet (#2108)
Closes #2107
<!--

Please add a reference to the issue that this PR addresses and indicate
which
files are most critical to review. If it fully addresses a particular
issue,
please include "Closes #XXX" (where "XXX" is the issue number).

If this PR is non-trivial/large/complex, please ensure that you have
either
created an issue that the team's had a chance to respond to, or had some
discussion with the team prior to submitting substantial pull requests.
The team
can be reached via GitHub Discussions or the Cosmos Network Discord
server in
the #cometbft channel. GitHub Discussions is preferred over Discord as
it
allows us to keep track of conversations topically.
https://github.com/cometbft/cometbft/discussions

If the work in this PR is not aligned with the team's current
priorities, please
be advised that it may take some time before it is merged - especially
if it has
not yet been discussed with the team.

See the project board for the team's current priorities:
https://github.com/orgs/cometbft/projects/1

-->
This PR introduces a new folder `test/monitoring` which contains a
Docker Compose file that creates a Grafana and a Prometheus server
locally for the developer. It connects up with a localnet installation.

The localnet Docker compose file (`compose.yml`) was extended to allow
Prometheus traffic from the container and documentation was updated to
reflect the changes.

Short way to execute:
```
make build-linux localnet-start monitoring-start
```
---

The JSON files are pre-populated Grafana dashboards. Two of them are
from the qa-infra repository, one is part of host monitoring with
"node-explorer" and one is a new one created for storage experiments
that are currently ongoing.

#### PR checklist

- [ ] Tests written/updated - no Go code change
- [X] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [X] Updated relevant documentation (`docs/` or `spec/`) and code
comments - test docs were updated
- [X] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec

---------

Co-authored-by: Anton Kaliaev <anton.kalyaev@gmail.com>
(cherry picked from commit e8b4efa1721deb9f4f40786cf13e05ec6d0bb36b)