test(mempool): fix TestMempoolUpdateDoesNotPanicWhenApplicationMissedTx (#2242)
The test `TestMempoolUpdateDoesNotPanicWhenApplicationMissedTx` has two
problems:
- First, the test calls `reqRes.InvokeCallback()` on every CheckTx
request-response, but there's no callback set on `reqRes` so the tx are
not really added to the mempool at this point.
- Near the end, it makes two calls to `callback` on requests of type
`CHECK_TX_TYPE_CHECK` when they should be of type
`CHECK_TX_TYPE_RECHECK`. The intention here was to process the recheck
responses; instead, the two txs are added to the mempool.